I don't vibe

+++ Assets/Scripts/Health/Health.cs
@@ -61,7 +61,7 @@ public abstract class Health : MonoBehaviourPunCallbacks, IDamageable, IPunInsta
     }
 
     [PunRPC]
-    public void RPC_UpdateHealthUI(int damage)
+    public void RPC_UpdateHealthUI(float damage)
     {
         healthItem.SetHealthUI(damage);
     }
